requirements.txt
annotated-types==0.7.0
anyio==4.9.0
astrapy==2.0.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
deprecation==2.1.0
dnspython==2.7.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
packaging==24.2
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.11.0
pydantic-core==2.33.0
pymongo==4.11.3
pyparsing==3.2.3
python-dotenv==1.1.0
openai==1.11.1
requests==2.32.3
rsa==4.9
sniffio==1.3.1
toml==0.10.2
tqdm==4.67.1
typing-extensions==4.13.0
typing-inspection==0.4.0
uritemplate==4.1.1
urllib3==2.3.0
uuid6==2024.7.10
